Lots of changes. More closely reconciled game model with controllers and views. Gave views access to stage. Created game loop. Working canvas in ship view.

+++ src/main/java/spacetrader/controlship/RealTimeShipController.java
@@ -39,7 +39,7 @@ class RealTimeShipController {
     public Player getPlayer(){
         return player;
     }
-    public void movePlayer(double x,double y){
+    public void movePlayer(float x,float y){
         player.getShip().getPosition().add(new Position(x,y));
     }
 }
